removing fails

+++ Content.IntegrationTests/Tests/Interaction/InteractionTest.Helpers.cs
@@ -125,7 +125,9 @@ public abstract partial class InteractionTest
             await Server.WaitPost(() =>
             {
                 Assert.That(HandSys.TryDrop(SEntMan.GetEntity(Player), null, false, true, Hands));
-                SEntMan.DeleteEntity(held);
+                if (SEntMan.EntityExists(held) && !SEntMan.Deleted(held))
+                    SEntMan.DeleteEntity(held);
+
                 SLogger.Debug($"Deleting held entity");
             });
         }
@@ -802,7 +804,13 @@ public abstract partial class InteractionTest
 
     protected async Task Delete(EntityUid uid)
     {
-        await Server.WaitPost(() => SEntMan.DeleteEntity(uid));
+        await Server.WaitPost(() =>
+        {
+            if (!SEntMan.EntityExists(uid) || SEntMan.Deleted(uid))
+                return;
+
+            SEntMan.DeleteEntity(uid);
+        });
         await RunTicks(5);
     }
